refactor(protocols): dedupe probePort, remove LW3 debug logs, worker-pool scan
Small-scope audit cleanups for the Electron main process protocol layer.

Dedupe
- probePort was an 18-line identical helper in both auto-detect.ts and
  network-scanner.ts. Extract to protocols/net-utils.ts and import from
  both. Magic literals 6107/9990/80 replaced with named constants at the
  call sites (LIGHTWARE_PORT, VIDEOHUB_PORT, KUMO_HTTP_PORT).

Logging
- Drop four leftover `console.log('[LW3] ...')` debug statements in
  lightware.ts (ProductName/label response dumps, connect result log,
  download preview). These were noisy in DevTools and never gated by a
  debug flag.

Perf
- network-scanner.scanSubnet previously ran 11 sequential batches of 25
  probes (BATCH_SIZE=25, 254/25 rounds). Replace with a worker-pool of 50
  concurrent probes pulling from a shared counter — roughly halves scan
  wall-time on a quiet /24. Also emits progress after every probe rather
  than only at batch boundaries.
