+++ electron/src/main/protocols/auto-detect.ts
@@ -3,42 +3,17 @@
 // 2. Videohub (TCP 9990)
 // 3. KUMO (HTTP 80)
 
-import * as net from 'net'
 import { RouterType } from './types'
 import { kumoTestConnection } from './kumo-rest'
+import { probePort } from './net-utils'
 
 const PROBE_TIMEOUT = 2500
-
-function probePort(ip: string, port: number, timeout = PROBE_TIMEOUT): Promise<boolean> {
-  return new Promise((resolve) => {
-    const sock = new net.Socket()
-    const timer = setTimeout(() => {
-      sock.destroy()
-      resolve(false)
-    }, timeout)
-
-    sock.connect(port, ip, () => {
-      clearTimeout(timer)
-      sock.destroy()
-      resolve(true)
-    })
-
-    sock.on('error', () => {
-      clearTimeout(timer)
-      resolve(false)
-    })
-  })
-}
+const LIGHTWARE_PORT = 6107
+const VIDEOHUB_PORT = 9990
 
 export async function detectRouterType(ip: string): Promise<RouterType | null> {
-  // Probe Lightware (6107)
-  if (await probePort(ip, 6107)) return 'lightware'
-
-  // Probe Videohub (9990)
-  if (await probePort(ip, 9990)) return 'videohub'
-
-  // Probe KUMO (HTTP 80 REST API)
+  if (await probePort(ip, LIGHTWARE_PORT, PROBE_TIMEOUT)) return 'lightware'
+  if (await probePort(ip, VIDEOHUB_PORT, PROBE_TIMEOUT)) return 'videohub'
   if (await kumoTestConnection(ip)) return 'kumo'
-
   return null
 }
